Add urls related record on Brand & fix unique slug
