+++ packages/admin/src/Filament/Resources/CollectionResource/Pages/ManageCollectionUrls.php
@@ -2,108 +2,13 @@
 
 namespace Lunar\Admin\Filament\Resources\CollectionResource\Pages;
 
-use Filament\Forms;
-use Filament\Forms\Form;
-use Filament\Resources\Pages\ManageRelatedRecords;
-use Filament\Support\Facades\FilamentIcon;
-use Filament\Tables;
-use Filament\Tables\Table;
-use Illuminate\Contracts\Support\Htmlable;
-use Illuminate\Database\Eloquent\Builder;
-use Illuminate\Support\Str;
-use Illuminate\Validation\Rules\Unique;
 use Lunar\Admin\Filament\Resources\CollectionResource;
-use Lunar\Models\Product;
+use Lunar\Admin\Support\Resources\Pages\ManageUrlsRelatedRecords;
+use Lunar\Models\Collection;
 
-class ManageCollectionUrls extends ManageRelatedRecords
+class ManageCollectionUrls extends ManageUrlsRelatedRecords
 {
     protected static string $resource = CollectionResource::class;
 
-    protected static string $relationship = 'urls';
-
-    public function getTitle(): string|Htmlable
-    {
-        return __('lunarpanel::relationmanagers.urls.title_plural');
-    }
-
-    public static function getNavigationIcon(): ?string
-    {
-        return FilamentIcon::resolve('lunar::urls');
-    }
-
-    public static function getNavigationLabel(): string
-    {
-        return __('lunarpanel::relationmanagers.urls.title_plural');
-    }
-
-    public function form(Form $form): Form
-    {
-        return $form
-            ->schema([
-                Forms\Components\Toggle::make('default')->label(
-                    __('lunarpanel::relationmanagers.urls.form.default.label')
-                )->columnSpan(2),
-                Forms\Components\Group::make([
-                    Forms\Components\TextInput::make('slug')
-                        ->label(
-                            __('lunarpanel::relationmanagers.urls.table.slug.label')
-                        )
-                        ->required()
-                        ->dehydrateStateUsing(
-                            fn ($state) => Str::slug($state)
-                        )->unique(
-                            ignoreRecord: true,
-                            modifyRuleUsing: function (Unique $rule) {
-                                return $rule->where('element_type', Product::class);
-                            }
-                        )
-                        ->maxLength(255)
-                        ->required(),
-                    Forms\Components\Select::make('language_id')->label(
-                        __('lunarpanel::relationmanagers.urls.table.language.label')
-                    )->relationship(name: 'language', titleAttribute: 'name')->required(),
-                ])->columns(2)->columnSpan(2),
-            ]);
-    }
-
-    public function table(Table $table): Table
-    {
-        return $table
-            ->recordTitleAttribute('name')
-            ->modifyQueryUsing(fn (Builder $query) => $query->orderBy('language_id')->orderBy('default', 'desc'))
-            ->columns([
-                Tables\Columns\TextColumn::make('slug')->label(
-                    __('lunarpanel::relationmanagers.urls.table.slug.label')
-                ),
-                Tables\Columns\TextColumn::make('language.name')->label(
-                    __('lunarpanel::relationmanagers.urls.table.language.label')
-                ),
-                Tables\Columns\IconColumn::make('default')
-                    ->label(
-                        __('lunarpanel::relationmanagers.urls.table.default.label')
-                    )
-                    ->boolean(),
-            ])
-            ->filters([
-                Tables\Filters\SelectFilter::make('language_id')
-                    ->label(
-                        __('lunarpanel::relationmanagers.urls.filters.language_id.label')
-                    )
-                    ->relationship('language', 'name'),
-            ])
-            ->headerActions([
-                Tables\Actions\CreateAction::make()->label(
-                    __('lunarpanel::relationmanagers.urls.actions.create.label')
-                ),
-            ])
-            ->actions([
-                Tables\Actions\EditAction::make(),
-                Tables\Actions\DeleteAction::make(),
-            ])
-            ->bulkActions([
-                Tables\Actions\BulkActionGroup::make([
-                    Tables\Actions\DeleteBulkAction::make(),
-                ]),
-            ]);
-    }
+    protected static string $model = Collection::class;
 }
